# core/openapi.yaml
openapi: 3.0.3
info:
  title: Pen CMS API Specification
  description: |
    The contract-first API specification for Pen CMS, a Git-native flat-file CMS.

    Humans and agents share one capability vocabulary (string lists). `write:*` is
    create+update (never split create:/update:). `delete:*` is separate. `read` is
    monolithic. `publish:content` is content status (PATCH approve/publish) and is
    not host `publish` (deploy + Deploy Grant). Generic PUT of a post/page does not
    sniff status — that stays under write:posts / write:pages.

    Legacy agent scope `write` expands one-way to every write:*, every delete:*, and
    publish:content. It never implies host publish, users:manage, or manage:sites.

    Human admin routes under /api authorize from cookie `pen_jwt` or Bearer JWT →
    user YAML. Forgeable `pen_role` / `pen_user_id` cookies are not a root of trust.
  version: 0.1.0
servers:
  - url: /api/v1
paths:
...
  /authors/:
    get:
      summary: List site authors
      description: |
        List contributor bios for the active Content site (`content/sites/{site_id}/authors.yaml`).
        Site scope via `X-Pen-Site-Id` (human admin). Independent of CMS users / UserPublic.
        Bios are plain text. Not exposed as MCP tools.
      parameters:
        - $ref: '#/components/parameters/PenSiteId'
      responses:
        '200':
          description: Authors for the active site.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorsListResponse'
    post:
      summary: Create site author
      description: Create one author bio on the active site. Slug optional (derived from name when omitted). Avatar not required.
      parameters:
        - $ref: '#/components/parameters/PenSiteId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorCreate'
      responses:
        '201':
          description: Author created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '400':
          description: Validation error (e.g. duplicate slug).
  /authors/{slug}:
    get:
      summary: Get site author
      parameters:
        - $ref: '#/components/parameters/PenSiteId'
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Author record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '404':
          description: Author not found on this site.
    put:
      summary: Update site author
      description: Partial update. Slug is immutable (path param only).
      parameters:
        - $ref: '#/components/parameters/PenSiteId'
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorUpdate'
      responses:
        '200':
          description: Updated author.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '404':
          description: Author not found on this site.
    delete:
      summary: Delete site author
      description: Delete one author and best-effort remove their avatar file.
      parameters:
        - $ref: '#/components/parameters/PenSiteId'
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted.
        '404':
          description: Author not found on this site.
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Pen-API-Key
    CookieJwtAuth:
      type: apiKey
      in: cookie
      name: pen_jwt
      description: Human session JWT. Authorize from this cookie → YAML. Never trust pen_role / pen_user_id.
    BearerJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Human or agent JWT. Agent tokens use type=agent + scopes + site_id; they do not inherit the sponsor admin role.
  parameters:
    PenSiteId:
      name: X-Pen-Site-Id
      in: header
      required: false
      schema:
        type: string
        default: default
      description: >
        Active Content site id for human requests. Honored only if the actor
        has membership on that site (or YAML role admin). Unknown ids → 400.
        Non-member → 403 site_access_denied. Agent JWTs ignore this header and
        use the token site_id only.
...
  schemas:
...
    Author:
      type: object
      required:
        - slug
        - name
      properties:
        slug:
          type: string
        name:
          type: string
          description: "Display name (editor copies this into post frontmatter author: byline)."
        bio:
          type: string
          description: Plain text biography.
        website:
          type: string
        avatar:
          type: string
          nullable: true
          description: "Logical asset path e.g. images/authors/jane-doe.webp"
        email:
          type: string
        role:
          type: string
          description: Optional label (not RBAC).
        sort_order:
          type: integer
          default: 0
    AuthorCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        slug:
          type: string
          description: Optional; derived from name when omitted.
        bio:
          type: string
        website:
          type: string
        avatar:
          type: string
          nullable: true
        email:
          type: string
        role:
          type: string
        sort_order:
          type: integer
          nullable: true
    AuthorUpdate:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
        website:
          type: string
        avatar:
          type: string
          nullable: true
        email:
          type: string
        role:
          type: string
        sort_order:
          type: integer
    AuthorsListResponse:
      type: object
      required:
        - site_id
        - authors
      properties:
        site_id:
          type: string
        authors:
          type: array
          items:
            $ref: '#/components/schemas/Author'
...